build: add osiris metadata/entrypoint
Add enough metadata to allow building macOS bundles via cargo-osiris. This adds:

 - Library entrypoint as replacement for `src/main.rs`
 - Generic Osiris metadata that defines the application ID, so the produced
   bundle uses `Wonderous.app`, rather than `makepad_wonderous.app`.
 - macOS platform metadata, so the app is only compiled for the native
   architecture (cross-arch builds are broken in makepad upstream).
 - macOS archive metadata, so the app archives use a proper Apple AppID.

If `cargo-osiris` is available in PATH, you can now use:

    cargo-osiris build --platform macos
    cargo-osiris archive --platform macos --archive macos-pkg

This will produce:

    ./target/osiris/platform/macos/Wonderous.app/
    ./target/osiris/platform/macos/Wonderous.pkg

The `pkg` product can be submitted to TestFlight and should pass all verification,
if a registered Apple AppID is used, as well as an macOS Distribution certificate
with related provisioning-profile is used.

Signed-off-by: David Rheinsberg <david@readahead.eu>